Extract CloudCity changes to LoggingService
